// src/modules/Tesoreria/TiposCuenta/types/accountTypeTypes.ts
export interface AccountTypeDTO {
    id?: number
    name: string
    description?: string
    active: boolean
}

export interface AccountTypeFormDTO {
    id?: number
    name: string
    description?: string
    active: boolean
}

export interface SelectOptionDTO {
    id: number
    label: string
}